SST script temporary patch for overwriting keyring file during sst.

+++ scripts/wsrep_sst_xtrabackup-v2.sh
@@ -1862,7 +1862,7 @@ then
                         "is unencrypted. Enable encryption for SST traffic"
         wsrep_log_error "Line $LINENO"
         wsrep_log_error "****************************************************** "
-        exit 22
+        # KH: exit 22
     fi
 
     # Create the SST info file
@@ -2245,10 +2245,13 @@ then
         # with ever increasing number of files and achieve nothing.
         find $ib_home_dir $ib_log_dir $ib_undo_dir $DATA -mindepth 1  -regex $cpat  -prune  -o -exec rm -rfv {} 1>/dev/null \+
 
-        if [[ -r "$keyring_file_data" ]] || [[ -r "${keyring_file_data}.backup" ]];
-        then
-          wsrep_log_info "Cleaning the existing keyring file"
-          rm -f "$keyring_file_data" "${keyring_file_data}.backup"
+        # KH:
+        if [[ -z $transition_key ]]; then
+            if [[ -r "$keyring_file_data" ]] || [[ -r "${keyring_file_data}.backup" ]];
+            then
+            wsrep_log_info "Cleaning the existing keyring file"
+            rm -f "$keyring_file_data" "${keyring_file_data}.backup"
+            fi
         fi
 
         # Clean the binlog dir (if it's explicitly specified)
@@ -2404,9 +2407,10 @@ then
 
         # If we have a transition key, remove the existing keyring file data
         # it will be recreated in the move-back operation
-        if [[ -n $transition_key && -n $keyring_file_data ]]; then
-            rm -f "$keyring_file_data"
-        fi
+        # KH:
+#        if [[ -n $transition_key && -n $keyring_file_data ]]; then
+#            rm -f "$keyring_file_data"
+#        fi
 
         wsrep_log_info "Moving the backup to ${TDATA}"
